Working Disk read and write.

+++ include/sys/ahci.h
@@ -336,7 +336,7 @@ typedef volatile struct {
 
 
 void probe_port(hba_mem_t *abar);
-int disk_rw(hba_port_t * port, uint32_t startl, uint32_t starth, uint16_t count, uint64_t *buf, uint8_t rw);
+int disk_rw(hba_port_t * port, uint32_t startl, uint32_t starth, uint16_t count, uint8_t *buf, uint8_t rw);
 
 
 #endif
